Finish Logic
