+++ SynthShop.Infrastructure/Data/Repositories/OrderRepository.cs
@@ -17,149 +17,60 @@ namespace SynthShop.Infrastructure.Data.Repositories
 
 
 
-
-        public async Task<Order> CreateOrder(Guid basketId, Guid customerId)
-        {
-            int retryCount = 0;
-            const int maxRetries = 3;
-
-            while (retryCount < maxRetries)
-            {
-                try
-                {
-                    var basket = await _dbContext.Baskets
-                        .Include(b => b.Items)
-                        .ThenInclude(i => i.Product)
-                        .FirstOrDefaultAsync(b => b.BasketId == basketId);
-
-                    if (basket == null)
-                        throw new InvalidOperationException("Basket not found.");
-
-                    if (!basket.CustomerId.HasValue)
-                    {
-                        basket.CustomerId = customerId;
-                    }
-
-                    var availabilityIssues = new List<string>();
-                    foreach (var item in basket.Items)
-                    {
-                        if (item.Product.StockQuantity < item.Quantity)
-                        {
-                            availabilityIssues.Add($"Not enough stock for {item.Product.Name}. Requested: {item.Quantity}, Available: {item.Product.StockQuantity}");
-                        }
-                        else
-                        {
-                            item.Product.StockQuantity -= item.Quantity;
-                        }
-                    }
-
-                    if (availabilityIssues.Any())
-                    {
-                        throw new InvalidOperationException("There are issues with product availability: " + string.Join(", ", availabilityIssues));
-                    }
-
-                    var order = new Order
-                    {
-                        OrderID = Guid.NewGuid(),
-                        OrderDate = DateTime.UtcNow,
-                        CustomerID = basket.CustomerId.Value,
-                        Status = OrderStatus.Pending,
-                        OrderItems = basket.Items.Select(bi => new OrderItem
-                        {
-                            OrderItemID = Guid.NewGuid(),
-                            ProductID = bi.ProductId,
-                            Quantity = bi.Quantity,
-                            Price = bi.Product.Price,
-                            CreatedAt = DateTime.UtcNow
-                        }).ToList(),
-                        TotalAmount = basket.Items.Sum(i => i.Product.Price * i.Quantity),
-                        CreatedAt = DateTime.UtcNow
-                    };
-
-                    _dbContext.Orders.Add(order);
-                    await _dbContext.SaveChangesAsync();
-
-                    return order; 
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    
-                    if (retryCount == maxRetries - 1)
-                    {
-                        throw; 
-                    }
-                    retryCount++; 
-                }
-            }
-
-            throw new InvalidOperationException("Failed to create order due to concurrency conflicts.");
-        }
-
-
-
-
-
-
-
-
-        public async Task CancelOrder(Guid orderId)
+        public async Task<Order?> GetOrderAsync(Guid orderId)
         {
             var order = await _dbContext.Orders
                 .Include(o => o.OrderItems)
                 .ThenInclude(oi => oi.Product)
                 .SingleOrDefaultAsync(o => o.OrderID == orderId);
-
-            if (order == null)
-            {
-                throw new InvalidOperationException("Order not found.");
-            }
-
-            if (order.Status == OrderStatus.Completed)
-            {
-                throw new InvalidOperationException("Completed orders cannot be cancelled.");
-            }
-
-
-            if (order.Status == OrderStatus.Pending)
-            {
-                foreach (var item in order.OrderItems)
-                {
-                    var product = item.Product;  
-                    product.StockQuantity += item.Quantity;
-                }
-            }
             
-            order.Status = OrderStatus.Cancelled;
+            return order;
+        }
+
+        public async Task<Order> CreateOrderAsync(Order order)
+        {
+            _dbContext.Orders.Add(order);
+            await _dbContext.SaveChangesAsync();
+            return order;
+        }
 
 
+
+
+
+
+
+
+        public async Task DeleteOrderAsync(Guid orderId)
+        {
+            var order = new Order() { OrderID = orderId, IsDeleted = true };
+            _dbContext.Orders.Attach(order);
+            _dbContext.Entry(order).Property(e => e.IsDeleted).IsModified = true;
             await _dbContext.SaveChangesAsync();
         }
 
 
-        public async Task CompleteOrder(Guid orderId)
+        public async Task<Order?> UpdateOrderAsync(Guid orderId, Order order)
         {
-            var order = await _dbContext.Orders
-                .Include(o => o.OrderItems)
-                .FirstOrDefaultAsync(o => o.OrderID == orderId);
-
-            if (order == null)
+            var existingOrder = await _dbContext.Orders.FirstOrDefaultAsync(x => x.OrderID == orderId);
+            if (existingOrder == null)
             {
-                throw new InvalidOperationException("Order not found.");
+                return null;
             }
 
-            if (order.Status == OrderStatus.Completed)
-            {
-                throw new InvalidOperationException("Order is already completed.");
-            }
-
-            if (order.Status == OrderStatus.Cancelled)
-            {
-                throw new InvalidOperationException("Cannot complete a cancelled order.");
-            }
-
-            order.Status = OrderStatus.Completed;
-
-            await _dbContext.SaveChangesAsync(); 
+            existingOrder.OrderDate = order.OrderDate;
+            existingOrder.CustomerID = order.CustomerID;
+            existingOrder.Status = order.Status;
+            existingOrder.TotalAmount = order.TotalAmount;
+            existingOrder.OrderItems = order.OrderItems;
+            existingOrder.UpdateAt = DateTime.UtcNow;
+            existingOrder.IsDeleted = order.IsDeleted;
+            
+            
+            
+            await _dbContext.SaveChangesAsync();
+            return order;
         }
+        
     }
 }
